InterestingFactCarusel was changed
